renamed IncidentTicketDetails to IncidentTicketDailyJob
